back/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
pydantic_core==2.27.2
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
starlette==0.46.0
requests==2.32.3
